feat: 상세 페이지 textarea 리팩토링
